ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-google-genai
openai
pydantic
python-dotenv
httpx
scikit-learn
numpy
